backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
httpx==0.27.2
cachetools==5.5.0

